refactor(options): remove `set_string_option_direct`
Ref: #25672
