+++ src/nvim/api/options.c
@@ -122,10 +122,10 @@ static buf_T *do_ft_buf(char *filetype, aco_save_T *aco, Error *err)
   aucmd_prepbuf(aco, ftbuf);
 
   TRY_WRAP(err, {
-    set_option_value("bufhidden", STATIC_CSTR_AS_OPTVAL("hide"), OPT_LOCAL);
-    set_option_value("buftype", STATIC_CSTR_AS_OPTVAL("nofile"), OPT_LOCAL);
-    set_option_value("swapfile", BOOLEAN_OPTVAL(false), OPT_LOCAL);
-    set_option_value("modeline", BOOLEAN_OPTVAL(false), OPT_LOCAL);  // 'nomodeline'
+    set_option_value("bufhidden", STATIC_CSTR_AS_OPTVAL("hide"), OPT_LOCAL, 0, false);
+    set_option_value("buftype", STATIC_CSTR_AS_OPTVAL("nofile"), OPT_LOCAL, 0, false);
+    set_option_value("swapfile", BOOLEAN_OPTVAL(false), OPT_LOCAL, 0, false);
+    set_option_value("modeline", BOOLEAN_OPTVAL(false), OPT_LOCAL, 0, false);  // 'nomodeline'
 
     ftbuf->b_p_ft = xstrdup(filetype);
     do_filetype_autocmd(ftbuf, false);
@@ -254,7 +254,7 @@ void nvim_set_option_value(uint64_t channel_id, String name, Object value, Dict(
   });
 
   WITH_SCRIPT_CONTEXT(channel_id, {
-    set_option_value_for(name.data, optval, scope, req_scope, to, err);
+    set_option_value_for(name.data, optval, scope, 0, false, req_scope, to, err);
   });
 }
 
@@ -319,261 +319,4 @@ Dictionary nvim_get_option_info2(String name, Dict(option) *opts, Error *err)
   win_T *win = (req_scope == kOptReqWin) ? (win_T *)from : curwin;
 
   return get_vimoption(name, scope, buf, win, err);
-}
-
-/// Switch current context to get/set option value for window/buffer.
-///
-/// @param[out]  ctx        Current context. switchwin_T for window and aco_save_T for buffer.
-/// @param       req_scope  Requested option scope. See OptReqScope in option.h.
-/// @param[in]   from       Target buffer/window.
-/// @param[out]  err        Error message, if any.
-///
-/// @return  true if context was switched, false otherwise.
-static bool switch_option_context(void *const ctx, OptReqScope req_scope, void *const from,
-                                  Error *err)
-{
-  switch (req_scope) {
-  case kOptReqWin: {
-    win_T *const win = (win_T *)from;
-    switchwin_T *const switchwin = (switchwin_T *)ctx;
-
-    if (win == curwin) {
-      return false;
-    }
-
-    if (switch_win_noblock(switchwin, win, win_find_tabpage(win), true)
-        == FAIL) {
-      restore_win_noblock(switchwin, true);
-
-      if (try_end(err)) {
-        return false;
-      }
-      api_set_error(err, kErrorTypeException, "Problem while switching windows");
-      return false;
-    }
-    return true;
-  }
-  case kOptReqBuf: {
-    buf_T *const buf = (buf_T *)from;
-    aco_save_T *const aco = (aco_save_T *)ctx;
-
-    if (buf == curbuf) {
-      return false;
-    }
-    aucmd_prepbuf(aco, buf);
-    return true;
-  }
-  case kOptReqGlobal:
-    return false;
-  }
-  UNREACHABLE;
-}
-
-/// Restore context after getting/setting option for window/buffer. See switch_option_context() for
-/// params.
-static void restore_option_context(void *const ctx, OptReqScope req_scope)
-{
-  switch (req_scope) {
-  case kOptReqWin:
-    restore_win_noblock((switchwin_T *)ctx, true);
-    break;
-  case kOptReqBuf:
-    aucmd_restbuf((aco_save_T *)ctx);
-    break;
-  case kOptReqGlobal:
-    break;
-  }
-}
-
-/// Get attributes for an option.
-///
-/// @param  name  Option name.
-///
-/// @return  Option attributes.
-///          0 for hidden or unknown option.
-///          See SOPT_* in option_defs.h for other flags.
-int get_option_attrs(char *name)
-{
-  int opt_idx = findoption(name);
-
-  if (opt_idx < 0) {
-    return 0;
-  }
-
-  vimoption_T *opt = get_option(opt_idx);
-
-  if (is_tty_option(opt->fullname)) {
-    return SOPT_STRING | SOPT_GLOBAL;
-  }
-
-  // Hidden option
-  if (opt->var == NULL) {
-    return 0;
-  }
-
-  int attrs = 0;
-
-  if (opt->flags & P_BOOL) {
-    attrs |= SOPT_BOOL;
-  } else if (opt->flags & P_NUM) {
-    attrs |= SOPT_NUM;
-  } else if (opt->flags & P_STRING) {
-    attrs |= SOPT_STRING;
-  }
-
-  if (opt->indir == PV_NONE || (opt->indir & PV_BOTH)) {
-    attrs |= SOPT_GLOBAL;
-  }
-  if (opt->indir & PV_WIN) {
-    attrs |= SOPT_WIN;
-  } else if (opt->indir & PV_BUF) {
-    attrs |= SOPT_BUF;
-  }
-
-  return attrs;
-}
-
-/// Check if option has a value in the requested scope.
-///
-/// @param  name       Option name.
-/// @param  req_scope  Requested option scope. See OptReqScope in option.h.
-///
-/// @return  true if option has a value in the requested scope, false otherwise.
-static bool option_has_scope(char *name, OptReqScope req_scope)
-{
-  int opt_idx = findoption(name);
-
-  if (opt_idx < 0) {
-    return false;
-  }
-
-  vimoption_T *opt = get_option(opt_idx);
-
-  // Hidden option.
-  if (opt->var == NULL) {
-    return false;
-  }
-  // TTY option.
-  if (is_tty_option(opt->fullname)) {
-    return req_scope == kOptReqGlobal;
-  }
-
-  switch (req_scope) {
-  case kOptReqGlobal:
-    return opt->var != VAR_WIN;
-  case kOptReqBuf:
-    return opt->indir & PV_BUF;
-  case kOptReqWin:
-    return opt->indir & PV_WIN;
-  }
-  UNREACHABLE;
-}
-
-/// Get the option value in the requested scope.
-///
-/// @param       name       Option name.
-/// @param       req_scope  Requested option scope. See OptReqScope in option.h.
-/// @param[in]   from       Pointer to buffer or window for local option value.
-/// @param[out]  err        Error message, if any.
-///
-/// @return  Option value in the requested scope. Returns a Nil option value if option is not found,
-/// hidden or if it isn't present in the requested scope. (i.e. has no global, window-local or
-/// buffer-local value depending on opt_scope).
-OptVal get_option_value_strict(char *name, OptReqScope req_scope, void *from, Error *err)
-{
-  OptVal retv = NIL_OPTVAL;
-
-  if (!option_has_scope(name, req_scope)) {
-    return retv;
-  }
-  if (get_tty_option(name, &retv.data.string.data)) {
-    retv.type = kOptValTypeString;
-    return retv;
-  }
-
-  int opt_idx = findoption(name);
-  assert(opt_idx != 0);  // option_has_scope() already verifies if option name is valid.
-
-  vimoption_T *opt = get_option(opt_idx);
-  switchwin_T switchwin;
-  aco_save_T aco;
-  void *ctx = req_scope == kOptReqWin ? (void *)&switchwin
-                                    : (req_scope == kOptReqBuf ? (void *)&aco : NULL);
-  bool switched = switch_option_context(ctx, req_scope, from, err);
-  if (ERROR_SET(err)) {
-    return retv;
-  }
-
-  char *varp = get_varp_scope(opt, req_scope == kOptReqGlobal ? OPT_GLOBAL : OPT_LOCAL);
-  retv = optval_from_varp(opt_idx, varp);
-
-  if (switched) {
-    restore_option_context(ctx, req_scope);
-  }
-
-  return retv;
-}
-
-/// Get option value for buffer / window.
-///
-/// @param[in]   name       Option name.
-/// @param[out]  flagsp     Set to the option flags (P_xxxx) (if not NULL).
-/// @param[in]   scope      Option scope (can be OPT_LOCAL, OPT_GLOBAL or a combination).
-/// @param[out]  hidden     Whether option is hidden.
-/// @param       req_scope  Requested option scope. See OptReqScope in option.h.
-/// @param[in]   from       Target buffer/window.
-/// @param[out]  err        Error message, if any.
-///
-/// @return  Option value. Must be freed by caller.
-OptVal get_option_value_for(const char *const name, uint32_t *flagsp, int scope, bool *hidden,
-                            const OptReqScope req_scope, void *const from, Error *err)
-{
-  switchwin_T switchwin;
-  aco_save_T aco;
-  void *ctx = req_scope == kOptReqWin ? (void *)&switchwin
-                                      : (req_scope == kOptReqBuf ? (void *)&aco : NULL);
-
-  bool switched = switch_option_context(ctx, req_scope, from, err);
-  if (ERROR_SET(err)) {
-    return NIL_OPTVAL;
-  }
-
-  OptVal retv = get_option_value(name, flagsp, scope, hidden);
-
-  if (switched) {
-    restore_option_context(ctx, req_scope);
-  }
-
-  return retv;
-}
-
-/// Set option value for buffer / window.
-///
-/// @param[in]   name        Option name.
-/// @param[in]   value       Option value.
-/// @param[in]   opt_flags   Flags: OPT_LOCAL, OPT_GLOBAL, or 0 (both).
-/// @param       req_scope   Requested option scope. See OptReqScope in option.h.
-/// @param[in]   from        Target buffer/window.
-/// @param[out]  err         Error message, if any.
-void set_option_value_for(const char *const name, OptVal value, const int opt_flags,
-                          const OptReqScope req_scope, void *const from, Error *err)
-{
-  switchwin_T switchwin;
-  aco_save_T aco;
-  void *ctx = req_scope == kOptReqWin ? (void *)&switchwin
-                                      : (req_scope == kOptReqBuf ? (void *)&aco : NULL);
-
-  bool switched = switch_option_context(ctx, req_scope, from, err);
-  if (ERROR_SET(err)) {
-    return;
-  }
-
-  const char *const errmsg = set_option_value(name, value, opt_flags);
-  if (errmsg) {
-    api_set_error(err, kErrorTypeException, "%s", errmsg);
-  }
-
-  if (switched) {
-    restore_option_context(ctx, req_scope);
-  }
 }
