+++ src/nvim/ex_docmd.c
@@ -2654,7 +2654,7 @@ static void apply_cmdmod(cmdmod_T *cmod)
     // Set 'eventignore' to "all".
     // First save the existing option value for restoring it later.
     cmod->cmod_save_ei = xstrdup(p_ei);
-    set_string_option_direct("ei", -1, "all", OPT_FREE, SID_NONE);
+    set_option_value("ei", STATIC_CSTR_AS_OPTVAL("all"), OPT_FREE, SID_NONE, true);
   }
 }
 
@@ -2674,7 +2674,7 @@ void undo_cmdmod(cmdmod_T *cmod)
 
   if (cmod->cmod_save_ei != NULL) {
     // Restore 'eventignore' to the value before ":noautocmd".
-    set_string_option_direct("ei", -1, cmod->cmod_save_ei, OPT_FREE, SID_NONE);
+    set_option_value("ei", CSTR_AS_OPTVAL(cmod->cmod_save_ei), OPT_FREE, SID_NONE, true);
     free_string_option(cmod->cmod_save_ei);
     cmod->cmod_save_ei = NULL;
   }
