+++ src/nvim/fileio.c
@@ -1616,7 +1616,7 @@ failed:
     save_file_ff(curbuf);
     // If editing a new file: set 'fenc' for the current buffer.
     // Also for ":read ++edit file".
-    set_string_option_direct("fenc", -1, fenc, OPT_FREE | OPT_LOCAL, 0);
+    set_option_value("fenc", CSTR_AS_OPTVAL(fenc), OPT_FREE | OPT_LOCAL, 0, true);
   }
   if (fenc_alloced) {
     xfree(fenc);
@@ -1967,7 +1967,7 @@ void set_forced_fenc(exarg_T *eap)
   }
 
   char *fenc = enc_canonize(eap->cmd + eap->force_enc);
-  set_string_option_direct("fenc", -1, fenc, OPT_FREE|OPT_LOCAL, 0);
+  set_option_value("fenc", CSTR_AS_OPTVAL(fenc), OPT_FREE|OPT_LOCAL, 0, true);
   xfree(fenc);
 }
