+++ src/nvim/indent.c
@@ -1094,7 +1094,7 @@ void ex_retab(exarg_T *eap)
     colnr_T *old_vts_ary = curbuf->b_p_vts_array;
 
     if (tabstop_count(old_vts_ary) > 0 || tabstop_count(new_vts_array) > 1) {
-      set_string_option_direct("vts", -1, new_ts_str, OPT_FREE | OPT_LOCAL, 0);
+      set_option_value("vts", CSTR_AS_OPTVAL(new_ts_str), OPT_FREE | OPT_LOCAL, 0, true);
       curbuf->b_p_vts_array = new_vts_array;
       xfree(old_vts_ary);
     } else {
